+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "lint": "standard --fix",
     "rebuild": "cross-script npm rebuild --runtime=electron \"--target=$(electron -v)\" \"--abi=$(electron --abi)\" --disturl=https://atom.io/download/atom-shell",
-    "start": "electron index.js -- --title patchbay",
+    "start": "electron ahoy.js && electron index.js -- --title patchbay",
     "start-frameless": "FRAME=false npm start",
     "dev": "echo 'run your own sbot!' && electro main.js -- --title patchbay --icon ./assets/icon.png",
     "postinstall": "npm run rebuild",
@@ -89,14 +89,14 @@
     "scuttle-thread": "^1.0.1",
     "setimmediate": "^1.0.5",
     "ssb-about": "^2.0.0",
-    "ssb-ahoy": "0.0.1",
+    "ssb-ahoy": "^0.2.0",
     "ssb-backlinks": "^0.7.3",
     "ssb-blob-files": "^1.1.3",
     "ssb-blobs": "^1.1.12",
     "ssb-chess-db": "^1.0.5",
     "ssb-chess-mithril": "1.0.10",
     "ssb-config": "^3.2.3",
-    "ssb-ebt": "^5.3.11",
+    "ssb-ebt": "^5.4.3",
     "ssb-friend-pub": "^1.0.4",
     "ssb-friends": "^3.1.12",
     "ssb-gossip": "^1.0.5",
